requirements.txt
pandas>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
openpyxl>=3.1  # Optional: only needed if you read Excel files
